+++ src/main/scala/hm/moe/pokkedoll/warscore/games/TeamDeathMatch4.scala
@@ -639,7 +639,7 @@ class TeamDeathMatch4(override val id: String) extends Game {
 
     comp.append("* ")
       .append("K/D: ").color(ChatColor.GRAY)
-      .append((data.kill / (data.death + 1)).toString).color(ChatColor.GREEN).bold(true)
+      .append((data.kill / (data.death + 1).toDouble).toString).color(ChatColor.GREEN).bold(true)
       .append("\n").color(ChatColor.RESET).bold(false)
     /*
         comp.append("* ")
